refactor: season 아키텍처 개선
