+++ src/main/java/gravit/code/season/calendar/dto/SeasonDto.java
@@ -2,7 +2,7 @@ package gravit.code.season.calendar.dto;
 
 import java.time.LocalDateTime;
 
-public record SeasonResponse(
+public record SeasonDto(
         String seasonKey,
         LocalDateTime startsAt,
         LocalDateTime endsAt
