+++ src/test/java/gravit/code/season/batch/SeasonBatchServiceTest.java
@@ -5,7 +5,7 @@
 //import gravit.code.league.repository.LeagueRepository;
 //import gravit.code.season.domain.Season;
 //import gravit.code.season.domain.SeasonStatus;
-//import gravit.code.season.infrastructure.SeasonRepository;
+//import gravit.code.season.repository.SeasonRepository;
 //import gravit.code.season.service.port.SeasonClosedCache;
 //import gravit.code.userLeague.domain.UserLeagueRepository;
 //import gravit.code.userLeagueHistory.infrastructure.UserLeagueHistoryRepository;
