+++ src/test/java/gravit/code/user/domain/UserRepositoryTest.java
@@ -13,7 +13,7 @@
 //import gravit.code.progress.domain.ChapterProgress;
 //import gravit.code.progress.infrastructure.ChapterProgressJpaRepository;
 //import gravit.code.season.domain.Season;
-//import gravit.code.season.infrastructure.SeasonRepository;
+//import gravit.code.season.repository.SeasonRepository;
 //import gravit.code.user.infrastructure.UserJpaRepository;
 //import gravit.code.userLeague.domain.UserLeague;
 //import gravit.code.userLeague.infrastructure.UserLeagueJpaRepository;
